refactor: refactor BucketCreatePage (#27)
